refactor: use from for datafusion error
