+++ src/common/infra/errors.rs
@@ -35,8 +35,6 @@ pub enum Error {
     SimdJsonError(#[from] simd_json::Error),
     #[error("ArrowError# {0}")]
     ArrowError(#[from] datafusion::arrow::error::ArrowError),
-    #[error("DataFusionError# {0}")]
-    DataFusionError(#[from] datafusion::error::DataFusionError),
     #[error("WatchError# watcher is exists {0}")]
     WatcherExists(String),
     #[error("StringUTF8Error# {0}")]
